Update method changing, typo corecting, Table Class changing

+++ src/main/resources/init_db.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `books`.`books` (
     `id` INT NOT NULL AUTO_INCREMENT,
     `title` VARCHAR(255) NOT NULL,
-    `price` BIGINT NOT NULL,
+    `price` DECIMAL NOT NULL,
     PRIMARY KEY (`id`));
 
 INSERT INTO books (title, price) VALUES('First Book', 230),
